introduce an opcode dispatch mechanism
The previous way used to write optimizations was messy and complicated.
It required writing 'extra_' stencils, and using them by writing new code in the compiler itself.
This was not a good solution, it was fragile at best, and definitely not going to make it for the
future optimizations I have in mind.
This commit introduces a new dispatch mechanism.
A bunch of macros in stencils.c will generate some C strings containing pieces of code, along with
the stencil selection for a given opcode.
The stencil-builder.py script will collect these pieces and glue them together in a opcode-specific
dispatch function.
And that's all there is to it.

I still have to convert the remaining opcodes to using this new macro system, and this first version
is a bit slower than I wish from a compiler point of view: for each step, a check for the presence
of a dispatcher function is done. If the function exists, it is called and the opcode number is
replaced.
This works, but next step will just be to have a single huge dispatch function, automatically
generated of course, that will have a switch (opcode) and then each dispatch conditions.
It should be more efficient than this first version, but I can not be sure without some measurements
and I want this first version to be pushed and completed before changing more stuff.